# Conflicts:
#   raf-service/src/main/java/com/ozguryazilim/raf/RafService.java
#   raf-ui/src/main/resources/RafUiModule-messages.properties
#   raf-ui/src/main/resources/RafUiModule-messages_tr.properties

+++ raf-ui/src/main/java/com/ozguryazilim/raf/member/RafMemberController.java
@@ -19,7 +19,6 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.Locale;
-import java.util.logging.Level;
 import java.util.stream.Collectors;
 import javax.enterprise.event.Event;
 import javax.inject.Inject;
@@ -255,9 +254,22 @@ public class RafMemberController implements Serializable {
         return memberService.getGroupUsers(userGroupName).stream().map(m -> userLookup.getUserName(m)).collect(Collectors.toList());
     }
 
+    private boolean isLastManagerMember(RafMember member) {
+        if (RafDefinitionService.RAF_ROLE_MANAGER.equals(member.getRole())) {
+            List<RafMember> managerMembers = getManagerMembers();
+            return managerMembers.isEmpty() || (managerMembers.size() == 1 && managerMembers.get(0) == member);
+        }
+        return false;
+    }
+
     public void deleteMember(RafMember member) {
         try {
-            memberService.removeMember(member);
+            if (isLastManagerMember(member)) {
+                LOG.error("Last manager member cannot delete");
+                FacesMessages.error("Son Yönetici Üye Silinemez.", "Yeni bir yönetici üye ekledikten sonra tekrar deneyiniz.");
+            } else {
+                memberService.removeMember(member);
+            }
         } catch (RafException ex) {
             //FIXME: i18n
             LOG.error("Member cannot delete", ex);
@@ -308,5 +320,22 @@ public class RafMemberController implements Serializable {
     public void cancelDialog() {
         //Aslında yapacak bir şey yok.
         selectedMember = null;
+    }
+
+    public List<RafMember> getManagerMembers() {
+        try {
+            List<RafMember> result = memberService.getMembers(rafDefinition).stream()
+                    .filter(m
+                            -> m.getRole().equals(RafDefinitionService.RAF_ROLE_MANAGER))
+                    .collect(Collectors.toList());
+
+            return result;
+        } catch (RafException ex) {
+            //FIXME: i18n
+            LOG.error("Raf Exception", ex);
+            FacesMessages.error("Yönetici Üye bilgisi alınamadı", ex.getLocalizedMessage());
+        }
+
+        return Collections.emptyList();
     }
 }
